Refact: PV map forecast Add map loading state Add map failure state Change opacity function to linear Round expectedPowerGenerationMegawatts values Add zoom control to map component
