+++ apps/nowcasting-app/components/map/index.ts
@@ -1,0 +1,5 @@
+export { default as FaildStateMap } from "./faildState";
+export { default as Map } from "./map";
+export { default as LoadStateMap } from "./loadingState";
+export {default as PvLatestMap } from "./pvLatest";
+
