+++ apps/nowcasting-app/constant.ts
@@ -1,1 +1,2 @@
 export const API_PREFIX = "http://nowcasting-api-development.eu-west-1.elasticbeanstalk.com/v0";
+export const maxPowerGenerated = 500;
